// src/interfaces/CardProps.ts
export interface CardProps {
  backgroundColor: string;
  text1: string;
  text2: string;
  textColor1: string;
  textColor2: string;
  imgContent?: React.ReactNode;
  text3?: string;
  text4?: string;
  text5?: string;
  text6?: string;
  extraSvgContent?: React.ReactNode;
  height?: number;
  pr?: number;
  pb?: number;
  pl?: number;
  gap?: number;
  width?: string;
}